RanDAGen.jl Documentation

RanDAGen.jl is simple package that randomly generates directed acyclical graph's to be used in a variety of applications.

RanDAGen.Statistics

Base.getproperty

RanDAGen.print_stats

RanDAGen.Statistics — Type

AccumulatedStats{T<:Real} <: Stats{T}

AccumulatedStats()
AccumulatedStats(value::T)

Collection of statistics to describe the DAG.

Fields

min::T - the minimum for an array of values

max::T - the maximum for an array of values

n::Float64 - the weighted number of an array of values

sum::Float64 - the weighted sum for an array of values

sum_squares::Float64 - the weighted sum of squares for an array of values

Base.getproperty — Method

Overloaded statistics properties for mean, variance, and stddev.

RanDAGen.print_stats — Method

Print the statistics with optional format, title, and output stream

<!– @contents Pages = [ "index.md", ] –>